moving the js code to own file
