+++ lib/renderPage.js
@@ -1,5 +1,6 @@
 import { marked } from 'marked';
-import { loadMovie, loadMovies, loadMovieReviews } from '../src/apiMovies.js';
+import { loadMovie, loadMovies } from '../src/apiMovies.js';
+import { loadMovieReviews } from '../src/reviews.js';
 
 const MENU = [
     {
